// src/modules/email-templates/enum/email-templates-cases.enum.ts
export enum EmailTemplatesCases {
    APPOINTMENT_RECIVED = 'APPOINTMENT_RECIVED',
    SCHEDULE_APPOINTMENT_API = 'SCHEDULE_APPOINTMENT_API',
    CONFIRM_APPOINTMENT_API_TO_OWNER = 'CONFIRM_APPOINTMENT_API_TO_OWNER',
    APPOINTMENT_REJECTED_API_TO_CLIENT = 'APPOINTMENT_REJECTED_API_TO_CLIENT',
    CANCEL_OWNER_APPOINTMENT = 'CANCEL_OWNER_APPOINTMENT',
    APPOINTMENT_ACCEPTED = 'APPOINTMENT_ACCEPTED',
    APPOINTMENT_REMINDER = 'APPOINTMENT_REMINDER',
    APPOINTMENT_REMINDER_AGENT_OWNER = 'APPOINTMENT_REMINDER_AGENT_OWNER',
    APPOINTMENT_CLIENT_CANCELED = 'APPOINTMENT_CLIENT_CANCELED',
    RESCHEDULE_APPOINTMENT = 'RESCHEDULE_APPOINTMENT',
}